Rename ResChannel to ResponseChannel
