+++ atat/src/response_channel.rs
@@ -5,11 +5,11 @@ use embassy_sync::{
 
 use crate::Response;
 
-pub type ResChannel<const INGRESS_BUF_SIZE: usize> =
+pub type ResponseChannel<const INGRESS_BUF_SIZE: usize> =
     PubSubChannel<CriticalSectionRawMutex, Response<INGRESS_BUF_SIZE>, 1, 1, 1>;
 
-pub type ResPublisher<'sub, const INGRESS_BUF_SIZE: usize> =
+pub type ResponsePublisher<'sub, const INGRESS_BUF_SIZE: usize> =
     Publisher<'sub, CriticalSectionRawMutex, Response<INGRESS_BUF_SIZE>, 1, 1, 1>;
 
-pub type ResSubscription<'sub, const INGRESS_BUF_SIZE: usize> =
+pub type ResponseSubscription<'sub, const INGRESS_BUF_SIZE: usize> =
     Subscriber<'sub, CriticalSectionRawMutex, Response<INGRESS_BUF_SIZE>, 1, 1, 1>;
